third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/update.html

<!doctype html>
<meta charset="utf-8">
<title>StylePropertyMap.update tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#update-a-value-in-a-stylepropertymap">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/testhelper.js"></script>
<script>
'use strict';

const gInvalidTestCases = [
  { property: 'lemon', value: 'ade', desc: 'an unsupported property name' },
  { property: null, value: 'foo', desc: 'an null property name' },
  { property: 'width', value: CSS.deg(0), desc: 'an invalid CSSStyleValue' },
  { property: 'width', value: '10s', desc: 'an invalid String' },
];

for (const {property, value, desc} of gInvalidTestCases) {
  test(() => {
    let styleMap = newDivWithStyle('').attributeStyleMap;
    assert_throws(new TypeError(), () => styleMap.update(property, () => value));
  }, 'Updating a StylePropertyMap with ' + desc + ' throws TypeError');
}

test(() => {
  let styleMap = newDivWithStyle().attributeStyleMap;

  styleMap.update('width', () => CSS.px(10));
  assert_style_value_array_equals(styleMap.get('width'), CSS.px(10));

  styleMap.update('width', () => '20px');
  assert_style_value_array_equals(styleMap.get('width'), CSS.px(20));
}, 'Updating a property with CSSStyleValue or String updates its value');

test(() => {
  let styleMap = newDivWithStyle().attributeStyleMap;

  styleMap.update('transition-duration', () => CSS.s(1));
  assert_style_value_array_equals(styleMap.getAll('transition-duration'), [CSS.s(1)]);

  styleMap.update('transition-duration', () => '2s');
  assert_style_value_array_equals(styleMap.getAll('transition-duration'), [CSS.s(2)]);
}, 'Updating a list-valued property with CSSStyleValue or String updates its value');

test(() => {
  let styleMap = newDivWithStyle().attributeStyleMap;

  styleMap.update('--foo', () => CSS.px(10));
  assert_style_value_array_equals(styleMap.get('--foo'), CSS.px(10));

  styleMap.update('--foo', () => '20px');
  assert_style_value_array_equals(styleMap.get('--foo'), new CSSUnparsedValue('20px'));
}, 'Updating a custom property with CSSStyleValue or String updates its value');

test(() => {
  let styleMap = newDivWithStyle('').attributeStyleMap;
  styleMap.update('width', oldValue => assert_equals(oldValue, null));
}, 'Calling StylePropertyMap.update on an empty property model calls update function with null');

test(() => {
  let styleMap = newDivWithStyle('width: 10px').attributeStyleMap;
  styleMap.update('width', oldValue => assert_style_value_equals(oldValue, CSS.px(10)));
}, 'Calling StylePropertyMap.update on an existing property calls update function with old value');

test(() => {
  let styleMap = newDivWithStyle('transition-duration: 1s, 2s').attributeStyleMap;
  styleMap.update('transition-duration', oldValue => assert_style_value_equals(oldValue, CSS.s(1)));
}, 'Calling StylePropertyMap.update on an existing list-valued property calls update function with first value');

test(() => {
  let styleMap = newDivWithStyle('transition-duration: 5s, 10s').attributeStyleMap;

  styleMap.update('tRaNsItIoN-dUrAtIoN', () => ['1s', CSS.s(2)]);
  const result = styleMap.getAll('transition-duration');
  assert_style_value_array_equals(result, [CSS.s(1), CSS.s(2)]);
}, 'StylePropertyMap.update is case-insensitive');

</script>
</script>
